Navmesh
Added a basis and slightly hackish navmesh using unity's system. Along
with a test agent that can be selected and told to move with the mouse
buttons. Did a major refactor on the wall connections prepping them to
add more wall types later. Every case for the connections must be in and
verified before more types are added just for ease of use later.